chore(SecurityHub): Add class
