+++ prowler/__main__.py
@@ -40,15 +40,10 @@ from prowler.lib.outputs.compliance.compliance import display_compliance_table
 from prowler.lib.outputs.html.html import add_html_footer, fill_html_overview_statistics
 from prowler.lib.outputs.json.json import close_json
 from prowler.lib.outputs.outputs import extract_findings_statistics
+from prowler.lib.outputs.security_hub.security_hub import SecurityHub
 from prowler.lib.outputs.slack.slack import Slack
 from prowler.lib.outputs.summary_table import display_summary_table
 from prowler.providers.aws.lib.s3.s3 import send_to_s3_bucket
-from prowler.providers.aws.lib.security_hub.security_hub import (
-    batch_send_to_security_hub,
-    prepare_security_hub_findings,
-    resolve_security_hub_previous_findings,
-    verify_security_hub_integration_enabled_per_region,
-)
 from prowler.providers.common.provider import Provider
 from prowler.providers.common.quick_inventory import run_provider_quick_inventory
 
@@ -320,42 +315,43 @@ def prowler():
             if not global_provider.identity.audited_regions
             else global_provider.identity.audited_regions
         )
+        security_hub = SecurityHub(global_provider)
+
         for region in security_hub_regions:
             # Save the regions where AWS Security Hub is enabled
-            if verify_security_hub_integration_enabled_per_region(
-                global_provider.identity.partition,
+            if security_hub.verify_security_hub_integration_enabled_per_region(
                 region,
-                global_provider.session.current_session,
-                global_provider.identity.account,
             ):
                 aws_security_enabled_regions.append(region)
 
         # Prepare the findings to be sent to Security Hub
-        security_hub_findings_per_region = prepare_security_hub_findings(
+        security_hub_findings_per_region = security_hub.prepare_security_hub_findings(
             findings,
-            global_provider,
-            global_provider.output_options,
             aws_security_enabled_regions,
         )
 
         # Send the findings to Security Hub
-        findings_sent_to_security_hub = batch_send_to_security_hub(
-            security_hub_findings_per_region, global_provider.session.current_session
+        findings_sent_to_security_hub = security_hub.batch_send_to_security_hub(
+            security_hub_findings_per_region
         )
 
+        # Refactor(CLI)
         print(
             f"{Style.BRIGHT}{Fore.GREEN}\n{findings_sent_to_security_hub} findings sent to AWS Security Hub!{Style.RESET_ALL}"
         )
 
         # Resolve previous fails of Security Hub
         if not args.skip_sh_update:
+            # Refactor(CLI)
             print(
                 f"{Style.BRIGHT}\nArchiving previous findings in AWS Security Hub, please wait...{Style.RESET_ALL}"
             )
-            findings_archived_in_security_hub = resolve_security_hub_previous_findings(
-                security_hub_findings_per_region,
-                global_provider,
+            findings_archived_in_security_hub = (
+                security_hub.resolve_security_hub_previous_findings(
+                    security_hub_findings_per_region,
+                )
             )
+            # Refactor(CLI)
             print(
                 f"{Style.BRIGHT}{Fore.GREEN}\n{findings_archived_in_security_hub} findings archived in AWS Security Hub!{Style.RESET_ALL}"
             )
