+++ tests/lib/outputs/security_hub/fixtures/metadata.json
